+++ src/integrationTests/util.ts
@@ -16,6 +16,29 @@ export type TestMultiassetsWithFormat = [
 	expected: `0x${string}`,
 ];
 
+export const paraTransferAssets = async <T extends Format>(
+	parachainATA: AssetTransferApi,
+	format: T,
+	xcmVersion: number,
+	destChainId: string,
+	assetIds: string[],
+	amounts: string[],
+	opts: CreateXcmCallOpts,
+): Promise<TxResult<T>> => {
+	return await parachainATA.createTransferTransaction(
+		destChainId,
+		'0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
+		assetIds,
+		amounts,
+		{
+			format,
+			xcmVersion,
+			weightLimit: opts.weightLimit,
+			sendersAddr: 'FBeL7DanUDs5SZrxZY1CizMaPgG9vZgJgvr52C2dg81SsF1',
+		},
+	);
+};
+
 export const paraTransferMultiasset = async <T extends Format>(
 	parachainATA: AssetTransferApi,
 	format: T,
@@ -34,6 +57,7 @@ export const paraTransferMultiasset = async <T extends Format>(
 			xcmVersion,
 			weightLimit: opts.weightLimit,
 			sendersAddr: 'FBeL7DanUDs5SZrxZY1CizMaPgG9vZgJgvr52C2dg81SsF1',
+			xcmPalletOverride: 'xTokens',
 		},
 	);
 };
@@ -58,6 +82,7 @@ export const paraTransferMultiassets = async <T extends Format>(
 			xcmVersion,
 			weightLimit: opts.weightLimit,
 			sendersAddr: 'FBeL7DanUDs5SZrxZY1CizMaPgG9vZgJgvr52C2dg81SsF1',
+			xcmPalletOverride: 'xTokens',
 		},
 	);
 };
@@ -82,27 +107,7 @@ export const paraTransferMultiassetWithFee = async <T extends Format>(
 			xcmVersion,
 			weightLimit: opts.weightLimit,
 			sendersAddr: 'FBeL7DanUDs5SZrxZY1CizMaPgG9vZgJgvr52C2dg81SsF1',
-		},
-	);
-};
-
-export const paraTeleportNativeAsset = async <T extends Format>(
-	parachainATA: AssetTransferApi,
-	format: T,
-	nativeAssetId: string,
-	xcmVersion: number,
-	opts: CreateXcmCallOpts,
-): Promise<TxResult<T>> => {
-	return await parachainATA.createTransferTransaction(
-		'1000', // `1000` indicating the dest chain is a system chain.
-		'FBeL7DanUDs5SZrxZY1CizMaPgG9vZgJgvr52C2dg81SsF1',
-		[nativeAssetId],
-		['10000000000'],
-		{
-			format,
-			xcmVersion,
-			weightLimit: opts.weightLimit,
-			sendersAddr: '0xf5d5714c084c112843aca74f8c498da06cc5a2d63153b825189baa51043b1f0b',
+			xcmPalletOverride: 'xTokens',
 		},
 	);
 };